refactor: background_utils -> _background_subtraction
