+++ pyxem/tests/utils/test_dask_tools.py
@@ -23,7 +23,7 @@ import skimage.morphology as sm
 
 from pyxem.signals import Diffraction2D, LazyDiffraction2D
 import pyxem.utils.dask_tools as dt
-import pyxem.utils.background_utils as bt
+import pyxem.utils._background_subtraction as bt
 import pyxem.utils.expt_utils as et
 import pyxem.utils.pixelated_stem_tools as pst
